requirements.txt
Django==3.2
celery==5.2.3
django-celery-results==2.2.0
django-environ==0.8.1
django-extensions==3.1.5
djangorestframework==3.13.1
uWSGI==2.0.20
dj-static==0.0.6
static==1.1.1
static3==0.7.0
drf-extra-fields==3.2.1
drf-yasg==1.20.0
django-json-widget==1.1.1
django-leaflet==0.28.2
redis==4.1.4
psycopg2==2.9.3
psycopg2-binary==2.9.3
django-lifecycle==0.9.6
django-cors-headers==3.11.0
geopy==1.20.0
postgis==1.0.4